1. create a package to store all aliyun related function 2. allow vc-manager to add controller based on the type of the master_provisioner
