+++ incubator/virtualcluster/pkg/controller/add_virtualcluster.go
@@ -22,5 +22,5 @@ import (
 
 func init() {
 	// AddToManagerFuncs is a list of functions to create controllers and add them to a manager.
-	AddToManagerFuncs = append(AddToManagerFuncs, virtualcluster.Add)
+	AddToManagerFuncs[VirtualclusterController] = virtualcluster.Add
 }
